updated again

+++ test/tests/can_fd_queueSet_test.c
@@ -75,7 +75,7 @@ static void rx_task(void *pvParameters) {
         Error_Handler();
     }
 
-    uint16_t id = 0;
+    uint32_t id = 0;
 
     FDCAN_RxHeaderTypeDef fdcan1_rx_header = {0};
     uint8_t fdcan1_rx_data[8] = {0};
